Rename and reorder QuestionMetaPresenter file
